cub seg

+++ Cub3d/cub3d/Libft/ft_longest_node_len.c
@@ -6,7 +6,7 @@
 /*   By: seyun <seyun@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/12 12:25:01 by seyun             #+#    #+#             */
-/*   Updated: 2021/11/29 23:02:17 by seyun            ###   ########.fr       */
+/*   Updated: 2021/12/02 17:00:07 by seyun            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,13 @@
 
 int	ft_longest_node_len(t_list *lst)
 {
-	size_t	max;
+	int	max;
 
 	max = -1;
 	while (lst)
 	{
-		if (max < ft_strlen((char *)lst->content))
-			max = ft_strlen((char *)lst->content);
+		if (max < (int)ft_strlen((char *)lst->content))
+			max = (int)ft_strlen((char *)lst->content);
 		lst = lst->next;
 	}
 	return (max);
